requirements.txt
json2html
flask
obs-websocket-py
requests
waitress
google-api-python-client